feat: delete note
