+++ src/modules/notes/http/routes/notes.routes.ts
@@ -1,6 +1,7 @@
 import { Router } from 'express';
 
 import { createNotesController } from '../controllers/createNotes.controllers';
+import { deleteNotesControllers } from '../controllers/deleteNotes.controllers';
 import { listNotesController } from '../controllers/listNotes.controllers';
 
 import { ensureAuthenticated } from '@/middlewares/ensureAuthenticated';
@@ -8,4 +9,5 @@ import { ensureAuthenticated } from '@/middlewares/ensureAuthenticated';
 const notesRoutes = Router();
 notesRoutes.post('/', ensureAuthenticated, createNotesController.handle);
 notesRoutes.get('/:id', ensureAuthenticated, listNotesController.handle);
+notesRoutes.delete('/:id', ensureAuthenticated, deleteNotesControllers.handle);
 export { notesRoutes as notes };
