+++ src/modules/notes/model/links.ts
@@ -1,0 +1,6 @@
+export interface ILinks {
+  id: string;
+  url: string;
+  note_id: string;
+  created_at?: string | Date;
+}
