+++ src/modules/notes/model/tags.ts
@@ -1,0 +1,6 @@
+export interface ITags {
+  id: string;
+  name: string;
+  note_id: string;
+  user_id: string;
+}
